# optimize sql.HashOf