--- a/sql/cache.go
+++ b/sql/cache.go
@@ -17,6 +17,7 @@ package sql
 import (
 	"fmt"
 	"runtime"
+	"sync"
 
 	"github.com/cespare/xxhash/v2"
 
@@ -25,7 +26,9 @@ import (
 
 // HashOf returns a hash of the given value to be used as key in a cache.
 func HashOf(v Row) (uint64, error) {
-	hash := xxhash.New()
+	hash := digestPool.Get().(*xxhash.Digest)
+	hash.Reset()
+	defer digestPool.Put(hash)
 	for i, x := range v {
 		if i > 0 {
 			// separate each value in the row with a nil byte
@@ -38,11 +41,17 @@ func HashOf(v Row) (uint64, error) {
 		// TODO: we don't have the type info necessary to appropriately encode the value of a string with a non-standard
 		//  collation, which means that two strings that differ only in their collations will hash to the same value.
 		//  See rowexec/grouping_key()
-		if _, err := hash.Write([]byte(fmt.Sprintf("%v,", x))); err != nil {
+		if _, err := fmt.Fprintf(hash, "%v,", x); err != nil {
 			return 0, err
 		}
 	}
 	return hash.Sum64(), nil
+}
+
+var digestPool = sync.Pool{
+	New: func() any {
+		return xxhash.New()
+	},
 }
 
 // ErrKeyNotFound is returned when the key could not be found in the cache.