* pool *xxhash.Digest objects
* use fmt.Fprintf to write to hash

benchmark stats

```
oos: linux
goarch: amd64
pkg: github.com/dolthub/go-mysql-server/sql
cpu: AMD Ryzen 9 7900 12-Core Processor
                  │     b1      │                 b2                  │
                  │   sec/op    │    sec/op     vs base               │
HashOf-24           79.65n ± 4%   70.86n ±  7%  -11.03% (p=0.002 n=6)
ParallelHashOf-24   10.47n ± 4%   11.85n ± 19%        ~ (p=0.368 n=6)
geomean             28.88n        28.98n         +0.32%

                  │     b1     │                   b2                   │
                  │    B/op    │    B/op     vs base                    │
HashOf-24           4.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
ParallelHashOf-24   4.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
geomean             4.000                    ?                      ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                  │     b1     │                   b2                   │
                  │ allocs/op  │ allocs/op   vs base                    │
HashOf-24           2.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
ParallelHashOf-24   2.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
geomean             2.000                    ?                      ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```