@reusee 2024-03-12
> The change seems fine, but I think the benchmarks might need to save the result to a global variable to avoid dead code elimination.

fixed. we're checking the error return value so the call will not be eliminated by the compiler.